Prevent matching unrelated images
Because regexes defined by Kayobe are not strict enough, Kolla can build
container images unrelated to the deployment. For example, the default
service configuration will build:

- centos-binary-barbican-keystone-listener (matches keystone)
- centos-binary-prometheus-memcached-exporter (matches memcached)

This commit makes regexes stricter to avoid unecessary builds. Note that
some images are still built unecessarily, such as nova-compute-ironic
when ironic is disabled or neutron-metadata-agent-ovn when ovn is
disabled.

Change-Id: I67f0e48ee03b150f6fe52b1a3e7e727b79207231