Ubuntu: Set MTU for veth in the network files
Configuring veth pair MTU in the systemd-networkd .netdev files
does not set the correct MTU for the peer and there isn't
configuration option in the [Peer] section to do so [1].
This results in an MTU mismatch with peer having the default
(1500) value.

This change moves the MTU configuration from the .netdev file
to the [Link] section of the corresponding veth pair .network files.
Same approach is already recommended in [1] for the tun and
tap devices.

[1] https://www.freedesktop.org/software/systemd/man/systemd.netdev.html

Story: 2009072
Task: 42882
Change-Id: I082e1b807497d17ae40ddf86e1473fbfcdf625e2