Set proxy option in early dnf invocation
If hosts need to use a proxy to install packages and python is absent,
the call to install the python package will hang, at least on CentOS.

Set proxy if it exists in dnf_config using a command-line option. Later
uses of dnf will be after dnf.conf has been updated.

Change-Id: Ic949577afd3ea07306cebf224637a08a1ecce0f5
Story: 2009274
Task: 43528