CI: always return host configure test results
Currently, the overcloud host configure test results are only returned
on success, when they are least useful. This is because the zuul_return
task is not reached when the test task fails. This change reorders the
tasks to fix the issue.

Change-Id: Ie572eda567782cc377767bd8d925145e2ad4bd61
(cherry picked from commit c2ae611ee89173bc0888b9f37d51b934dcb2ee49)