infra VMs: use wait_for rather than wait_for_connection
wait_for_connection requires a python interpreter to be available, which
may not be the case in some images (CentOS Stream cloud images, cirros
etc.). Instead, use wait_for to wait for the SSH port to open, then rely
on the bootstrapping process to install an interpreter. This is the same
method used for the seed VM.

This change also syncs the SSH wait timeout for infra VMs which the one
used for seed VM provisioning (360 seconds).

Change-Id: I758aff1d3ef714f1c8ef82d29dd2217734a9aae6
(cherry picked from commit b87d89aae39cceedd4f70b447cddf5ada615bd36)