Fix installation prefix detection
When multiple 'lib' components are present in the installation path,
Kayobe fails to detect the right installation prefix which causes
various commands to raise an error.

Rewrite logic to detect the last 'lib' component in path instead. This
should work as long as this function is not stored under a lib directory
inside the Kayobe source tree.

Change-Id: Ie9b15db6563546ede9ce9735292ec566d540432a
Story: 2009721
Task: 44109