ntp: Fix service mask when service doesn't exist
The ntp role attempts to mask NTP services that may conflict with
chrony. Currently we are seeing the following failure:

    TASK ntp : Mask alternative NTP clients to prevent conflicts
    Could not find the requested service systemd-timesyncd.service: host

The service_facts module shows that the service is not found:

    "systemd-timesyncd.service": {
        "name": "systemd-timesyncd.service",
        "source": "systemd",
        "state": "stopped",
        "status": "not-found"
    },

According to the Internet, this can happen if there are After/Before
dependencies in one service on another that does not exist.

This change fixes the handling of these not-found services to avoid the
error.

Note this patch also reduces reserved blocks on the root file system to
work around disk space issues in CI.

Story: 2009821
Task: 44401

Change-Id: I4b8a42704f2b0a145ee9dec433d91df67628cd9d