CI: Revert from UEFI to BIOS mode in TLS jobs
We keep hitting the OOM killer in the TLS job, and the switch to UEFI
mode in bare metal testing seems like the most likely culprit.

This change keeps in place most of the changes of
Id1b4e9775c834b8b97e086241ee8b247977225a2, but removes the override of
ironic_boot_mode in the TLS job.

Change-Id: Id35894be893183b0b84e6dacb81e1b72e4a828f9