Use jinja2.pass_context instead of contextfilter
The contextfilter decorator was deprecated in jinja2 3.0.0, and has been
dropped in 3.1.0. This results in the following warning, and failed
attempts to use filters:

    [WARNING]: Skipping plugin (networks.py) as it seems to be invalid:
    module 'jinja2' has no attribute 'contextfilter'

This change switches to use the pass_context decorator. The minimum
version of Jinja2 is raised to 3 to ensure pass_context is present.

This backport also includes some changes to address issues with image
builds in CI, caused by CentOS Scream.

1. disable IPA image builds in seed deploy jobs

IPA image builds will be split out into a separate job. For now, disable
them.

2. copy a local file instead of downloading Cirros

This is a follow up to Ibc46a18d00c4c51b67ea18f24a4dd138ef59a4ff, which
did not work for the seed upgrade job when backported to xena. This is
because Wallaby did not support overriding dib.yml for bifrost.

This change works around the issue by copying a local file, instead of
downloading Cirros.

Depends-On: https://review.opendev.org/c/openstack/kayobe/+/835187
Change-Id: If657bf5b0117812d3c53942464cc41cf86cc8ad5