ironic: revert to ironic's default drivers & interfaces
There's no good reason to provide our own defaults for Ironic driver and
interface configuration. It easily gets out of sync with changes in
Ironic. Let's remove our defaults, and let Ironic provide them.

Note that this changes several of the defaults, which are listed in the
release note. Users are still able to override them using the existing
variables in ironic.yml.

Change-Id: Ie27adca0029e41439720eaac49b24cf13fc8068d