Improve comments in networking config files
In networks.yml and inventory group_vars that become part of
kayobe-config, there are a lot of commented out variables for networks
that may or may not exist in a deployment. These are not always that
helpful, and may be actively harmful if the network names in a
deployment are different.

This change removes these variables, and replaces them with a short
explanation of what to add to the section.

Change-Id: I88abe2c515b60e1f1c4c7c073818ff3e47436e64