Allow overriding earlier use of NetworkManager
veths have to be in nmconnection format, as nm-settings-ifcfg-rh has
no support for reading veth network-scripts, as can be seen by the
lack of a table at [1] "Table 37. veth setting". Hence why we're not
using NetworkManager with old network-scripts.

But if you set interfaces_use_networkmanager to true, you can already
start using NM on rhel8 platforms if you so wish.

[1] https://networkmanager.dev/docs/api/latest/nm-settings-ifcfg-rh.html

Change-Id: Ic4dd537bee826ee89c009bf2c4c95781af19cbad