Generate local Kolla Ansible config in check mode
If running a command in check mode such as

    kayobe overcloud service deploy --check

Kayobe does not generate the local configuration for Kolla Ansible. This
can lead to an inaccurate result when comparing with the remote
configuration, if there are changes in kayobe-config.

For example:

* Run kayobe overcloud service deploy
* Change a file in etc/kayobe/kolla
* Run kayobe overcloud service deploy --check --diff

We would expect that the changed file results in a diff against the
remote config. However there is no diff displayed.

This change fixes the issue by always generating the local Kolla Ansible
config, even in check mode.

Change-Id: Ic1dd075076ea186b0928bba1a235605c0cd2ec71
Story: 2010526
Task: 47132