dev: Improve error checking for config check functions
Various functions in the development/testing scripts rely on 'kayobe
configuration dump' to extract the value of flags. If this command fails
for any reason, we should exit the script. Currently, some places we
continue and return 1, since we check the output against the string
'true'.

The to_bool helper function handles failure by checking for a valid
boolean output, so let's use that everywhere.

Change-Id: I3a5a43fef9c3d68d0db02be12b9f892c437e513d