Add option to skip kolla docker registry login
This change adds a variable that can be used to disable kolla docker
registry login attempts, even when the registry username and password
are set.

This is required for deployments using a non-standard containerised
registry deployed on the seed during the deploy-container step,
since it takes place after the registry login attempt.

Change-Id: Ie17ef9ce1147cbaec2e42db932c7d59293b49b1b