Fix MTU of NetworkManager bridge VLAN interfaces
We are seeing an issue on Rocky Linux 9 where VLAN interfaces on bridges
have their MTU set to MTU-1 after rebooting. This is likely related to
how MTU is set by NetworkManager for bridges: it first sets the MTU to
MTU-1 before setting it to the correct value [1].

We can work around this issue by not explicitly setting the MTU
configuration on the VLAN interface if it is the same as the parent
bridge. It will be automatically inherited from the parent.

[1] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/864872e9a8c694ee7472154bf45f869040310b71

Closes-Bug: #2039947
Change-Id: I23366f4de7842e7c2fe40e431fac76f26e9892de
(cherry picked from commit ee63b3253dedaf62cc321f5b29080d7e7c459a99)