CI: Run configuration dump against localhost
Since using the to_bool function in more places in
I3a5a43fef9c3d68d0db02be12b9f892c437e513d, we are now more strict about
the result of the variable dump. If there are no controllers in the
inventory, the result will not be a valid boolean and the to_bool
function will exit non-zero.

This change fixes the issue by running against localhost, which should
always be in the inventory.

Change-Id: Idcfd9d335f11f6c4d676033128d207f62b363ee9