Use dummy1 as bridge port instead of eth1
The MichaelRigart.interfaces role only creates an interface with the
dummy type if it is called dummy*, otherwise it is created as ethernet.

The default dev configuration was using eth1 for the bridge port, which
was causing issues on Rocky Linux 9 when following the automated setup
instructions. Switch to dummy1 to ensure a dummy port is created.

Depends-On: https://review.opendev.org/c/openstack/kayobe-config-dev/+/904044

Change-Id: I64b6180194fd94c691e5ca75b815b6cb3f1d7d3c