Fix wipe-disks role to work on util-linux >= 2.37
On newer versions of util-linux, the ``wipe_disks`` role was not
properly configured due to the fact that ``lsblk -J`` returns a slightly
different output than what was previously returned. Previously the
output would contain a key called ``mountpoint``, however this has been
changed to ``mountpoints``. This has now been fixed by looking at the
new key as well as the old key.

Closes-bug: #2051859
Change-Id: I2a59f4eb3ae27f37fda38e05924315887cb04c41