Fix Dell OS6 and Dell OS9 switch configuration
Previously the dellos6_config and dellos9_config modules accepted the
name of a template file as their 'src' parameter, or a string containing
the actual config. Now they only accept the config string. Kayobe passes
the name of a template file, resulting in a strange error from the
switch about an invalid command dellos6-config.j2 or dellos9-config.j2.

This change fixes the issue in the same way it was fixed for Dell OS10
switches, by using the template lookup plugin to evaluate the config
file template.

Change-Id: Ic14e2b75c785b4db03b5e659e5a2be2ee83eace3
Closes-Bug: #2061102