Fix container image build issues with letsencrypt
The kolla_enable_letsencrypt variable was undefined, which was causing
overcloud container image build to fail when the variable was not
defined in Kayobe configuration.

The bool filter was missing for kolla_enable_letsencrypt, which was
causing letsencrypt images to be built when kolla_enable_letsencrypt was
set to "false" or "no".

Change-Id: I234b7a0434ba3d533a66b766331fd489078fbcf8
(cherry picked from commit 9669462208e1785c2d64485d6bdc10cc75210d39)