Use explicit version of community.docker
This works around a bug in the ansible packaged version of
community.docker when using requests>=2.32. It is also more flexible if
we can control the versions.

Also fixes an issue where kolla collections would not be installed if
the same role existed in the kayobe ansible collections path.

This relies on the support added in kolla-ansible to use ansible-core,
see: https://review.opendev.org/c/openstack/kolla-ansible/+/896406.

Closes-Bug: #2072980
Change-Id: Id9b19f10bc8ac38927914fb9782892b7daf1be82