Use kayobe_container only when service is enabled
The overcloud-extras playbook called during overcloud service deploy
would always invoke the kayobe_container module, even when the service
(inspection_store or opensm) was disabled.

With podman, this has the side effect of attempting to pull the image
from the registry, even when `state` is `absent`.

Change behaviour to invoke kayobe_container only when the service is
enabled. This means that a manual container removal is necessary now to
disable these services.

Change-Id: Id3ebd6836ba8f3b2101b94637a89f97fb0eb59e1