Add custom_etc_hosts_entries variable
``custom_etc_hosts_entries`` is a new variable which provides a generic
way to add entries to ``/etc/hosts``. It differs from
``etc_hosts_hosts`` because it is empty by default and can be used to
add hosts that are not in the Kayobe inventory.

``custom_etc_hosts_entries`` is a dictionary where each key is a
hostname and each value is an IP address.

Change-Id: I74c8234f77115d6a478559a53ad86534eb12e5f4
Signed-off-by: Alex Welsh <alex@stackhpc.com>