Remove undocumented playbook for creating external network
This is not referenced in the documentation and is not used by any
Kayobe commands. It is also likely to be broken as it needs refactoring
to properly support the stackhpc.os-networks role. An attempt was made
to do this previously[1], but the variables in the refactoring do not
exist in the new role. This wasn't caught at the time as it is
unused in the CI.

These days we prefer to manage the networks in a separate repository
using either Ansible or OpenTofu.

[1] https://github.com/openstack/kayobe/commit/619ad3ebc83811af640d4f6d4294b428946d4bd1

Change-Id: Ie3b618471e8ed91b202867972897e364805d336e
Signed-off-by: Will Szumski <will@stackhpc.com>